+++ src/test/resources/no/skatteetaten/aurora/boober/service/openshift/OpenShiftCommandBuilderTest/dc-webleveranse.json
