+++ src/test/resources/no/skatteetaten/aurora/boober/service/openshift/ResourceMergerTest/openshift-objects/service-new.json
