+++ src/test/resources/samples/result/booberdev/aos-simple/route.json
@@ -18,6 +18,7 @@
       "kind": "Service",
       "name": "aos-simple"
     },
+    "host": "aos-simple-aos-booberdev.utv.paas.skead.no",
     "path": "/foo"
   }
 }
