+++ src/test/resources/samples/result/booberdev/console/route.json
@@ -14,6 +14,7 @@
     "to": {
       "kind": "Service",
       "name": "openshift-console-api"
-    }
+    },
+    "host": "openshift-console-api-aos-booberdev.utv.paas.skead.no"
   }
 }
