+++ src/test/resources/samples/result/booberdev/reference-web/route.json
@@ -14,6 +14,7 @@
     "to": {
       "kind": "Service",
       "name": "openshift-referanse-react"
-    }
+    },
+    "host": "openshift-referanse-react-aos-booberdev.utv.paas.skead.no"
   }
 }
