+++ src/test/resources/samples/result/booberdev/reference/route.json
@@ -14,6 +14,7 @@
     "to": {
       "kind": "Service",
       "name": "reference"
-    }
+    },
+    "host": "reference-aos-booberdev.utv.paas.skead.no"
   }
 }
