+++ src/test/resources/samples/result/release/aos-simple/route.json
@@ -18,6 +18,7 @@
       "kind": "Service",
       "name": "aos-simple"
     },
+    "host": "aos-simple-aos-release.utv.paas.skead.no",
     "path": "/foo"
   }
 }
