+++ src/test/resources/samples/result/secrettest/aos-simple/route.json
@@ -18,6 +18,7 @@
       "kind": "Service",
       "name": "aos-simple"
     },
+    "host": "aos-simple-aos-secrettest.utv.paas.skead.no",
     "path": "/foo"
   }
 }
